requirements.txt
python-dotenv
PyGithub
pg8000
sqlalchemy
google-cloud-sqlconnector[pg8000]
pandas
streamlit